add payment link for tickets

+++ controllers/event.controller.js
@@ -4,6 +4,7 @@ const User = require('../models/user.model')
 const Team = require('../models/team.model')
 const { successfulRes, unsuccessfulRes } = require('../lib/response')
 const { sendEmail } = require('../lib/email')
+const stripe = require('stripe')(process.env.STRIPE_SECRET)
 
 // * CONTROLLERS * //
 // Get all events based on organization
@@ -48,8 +49,17 @@ const getSingleEvent = async (req, res) => {
 // Creat event for an organization by admin
 const createEvent = async (req, res) => {
   // Get the information from the request body
-  const { name, location, date, description, teamNames, workerEmails, sport } =
-    req.body
+  const {
+    name,
+    location,
+    date,
+    description,
+    teamNames,
+    workerEmails,
+    sport,
+    quantity,
+    price: itemPrice,
+  } = req.body
 
   // Get the workers from there email address
   const workers = await User.find({ email: workerEmails })
@@ -59,6 +69,40 @@ const createEvent = async (req, res) => {
 
   // Get the organization and the author from the user
   const { org, userId } = req.user
+
+  let paymentLink = 'NA'
+  let paymentLinkID = 'NA'
+
+  if (itemPrice) {
+    const product = await stripe.products.create({
+      name: 'Tickets for ' + name,
+    })
+
+    // ! CREATE STRIPE PAYMENT LINK
+    // Create a new price point for our secret image product
+    const price = await stripe.prices.create({
+      currency: 'usd',
+      unit_amount: itemPrice * 100,
+      product: product.id,
+    })
+
+    // Create a new payment session with that price id
+    paymentLink = await stripe.paymentLinks.create({
+      line_items: [
+        {
+          price: price.id,
+          adjustable_quantity: {
+            enabled: true,
+            minimum: 1,
+            maximum: quantity,
+          },
+          quantity: 1,
+        },
+      ],
+    })
+
+    paymentLinkID = paymentLink.id
+  }
 
   // Create the event
   const newEvent = await Event.create({
@@ -70,7 +114,9 @@ const createEvent = async (req, res) => {
     teams,
     workers,
     org,
+    paymentLinkID: paymentLink.id,
     author: userId,
+    ticketLink: paymentLink.url,
   })
 
   // send back the created event
@@ -192,6 +238,19 @@ const deleteEvent = async (req, res) => {
     return unsuccessfulRes({ res })
   }
 
+  //find the event
+  const foundEvent = await Event.findOne({ _id: id })
+
+  // if no event, return error
+  if (!foundEvent) {
+    return unsuccessfulRes({ res, status: 404, msg: 'Event not found' })
+  }
+
+  // deactivate the events payment link
+  await stripe.paymentLinks.update(foundEvent.paymentLinkID, {
+    active: false,
+  })
+
   // Delete the event
   const deletedEvent = await Event.findOneAndDelete({ _id: id })
 
@@ -259,6 +318,9 @@ const sendEmailToWorkers = async (req, res) => {
   return successfulRes({ res, data: workersEmails })
 }
 
+// send ticket purchase emails
+const purchaseTickets = async (req, res) => {}
+
 // * EXPORTS * //
 module.exports = {
   getAllEvents,
@@ -267,5 +329,6 @@ module.exports = {
   updateEvent,
   deleteEvent,
   sendEmailToWorkers,
+  purchaseTickets,
   updateScore,
 }
